Double tap to wake:initial commit
When the screen is double tapped at suspend, emulate the power button
press to wake it up. I have tested this feature for sometime in my
Nexus7. Some modifications have been made to increase the reliability.
Three sysfs interface are added to control it:
1. doubletap_wake, a swith to enable and disable this feature
2. dt_delta_limit, the bound between the two taps (in pixels)
3. dt_timeout, the max. duration between the two taps (in ms)
Settings are added to the PerformanceControl apk for smdk4210-tab.
Note:Tab plus use a different touch screen. It is not yet tested.